someRefactoring
